move checkAreTypesValid to typeHelpers
